# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * M&H Fashion Dropshipping Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  nameAr: string;
  nameEn: string;
  descriptionAr?: string;
  descriptionEn?: string;
  price: number;
  originalPrice?: number;
  imageUrl: string;
  images?: string[];
  categoryId?: number;
  categoryNameAr?: string;
  categoryNameEn?: string;
  stock: number;
  featured: boolean;
  sizes?: string[];
  colors?: string[];
  createdAt: string;
}